# apps/dashboard/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "anvil local dashboard API",
    "version": "0.9.0-beta",
    "description": "Loopback-only, read-only dashboard data for one local workspace."
  },
  "servers": [
    {
      "url": "/",
      "description": "Current loopback server"
    }
  ],
  "paths": {
...
    "/api/v1/protection": {
      "get": {
        "operationId": "getProtectionOverview",
        "summary": "Current local protection evidence",
        "tags": ["Protection"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectionOverview"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "421": {
            "description": "Loopback Host required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Dashboard worker failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/protection/history": {
      "get": {
        "operationId": "getProtectionHistory",
        "summary": "Retained local gate history",
        "tags": ["Protection"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProtectionHistory"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "421": {
            "description": "Loopback Host required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Dashboard worker failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/patterns": {
      "get": {
        "operationId": "getPatternCatalogue",
        "summary": "Compiled anti-pattern catalogue",
        "tags": ["Patterns"],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatternCatalogue"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "421": {
            "description": "Loopback Host required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Dashboard worker failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans": {
      "get": {
        "operationId": "listPlans",
        "summary": "Indexed APS plans",
        "tags": ["Plans"],
        "responses": {
          "200": {
            "description": "Plan summaries",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PlanSummary"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "421": {
            "description": "Loopback Host required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Dashboard worker failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Plan data unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/plans/{id}": {
      "get": {
        "operationId": "getPlan",
        "summary": "Selected APS plan detail",
        "tags": ["Plans"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[A-Za-z0-9_-]+$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Plan detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlanDetail"
                }
              }
            }
          },
          "400": {
            "description": "Invalid plan identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Cross-origin request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Plan not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "421": {
            "description": "Loopback Host required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "500": {
            "description": "Dashboard worker failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Plan data unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/openapi.json": {
      "get": {
        "operationId": "getOpenApi",
        "summary": "This OpenAPI document",
        "tags": ["Contract"],
        "responses": {
          "200": {
            "description": "OpenAPI 3.1 document"
          },
          "403": {
            "description": "Cross-origin request rejected",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "421": {
            "description": "Loopback Host required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DataState": {
        "type": "string",
        "enum": ["complete", "partial", "unavailable"]
      },
      "ProtectionHistoryPoint": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "recorded_at",
          "score",
          "status",
          "status_label",
          "warning_count",
          "duration_seconds",
          "checks_run"
        ],
        "properties": {
          "recorded_at": {
            "type": "string",
            "format": "date-time"
          },
          "score": {
            "type": "number",
            "minimum": 0,
            "maximum": 100
          },
          "status": {
            "type": "string",
            "enum": ["pass", "warn", "fail"]
          },
          "status_label": {
            "type": "string"
          },
          "warning_count": {
            "type": "integer",
            "minimum": 0
          },
          "duration_seconds": {
            "type": ["string", "null"]
          },
          "checks_run": {
            "type": ["string", "null"]
          }
        }
      },
      "ProtectionHistoryRange": {
        "type": "object",
        "additionalProperties": false,
        "required": ["first_recorded_at", "last_recorded_at"],
        "properties": {
          "first_recorded_at": {
            "type": "string",
            "format": "date-time"
          },
          "last_recorded_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "ProtectionHistory": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "data_state",
          "source_message",
          "actual_range",
          "points",
          "gaps"
        ],
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "anvil.dashboard.protection-history.v1"
          },
          "data_state": {
            "$ref": "#/components/schemas/DataState"
          },
          "source_message": {
            "type": "string"
          },
          "actual_range": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProtectionHistoryRange"
              },
              {
                "type": "null"
              }
            ]
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProtectionHistoryPoint"
            }
          },
          "gaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataGap"
            }
          }
        }
      },
      "ProtectionClaim": {
        "type": "object",
        "required": ["schema_version", "worktree_state", "surfaces"],
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "anvil.protection-claim.v1"
          },
          "worktree_state": {
            "type": "string",
            "enum": [
              "unprotected",
              "warming",
              "pre-write-embedded",
              "pre-write-daemon",
              "save-time-only",
              "full",
              "degraded-protection",
              "cross-boundary-mixed",
              "multi-daemon-detected",
              "path-uncertain"
            ]
          },
          "surfaces": {
            "type": "array",
            "items": {
              "type": "object",
              "required": ["identifier", "state"],
              "properties": {
                "identifier": {
                  "type": "string"
                },
                "state": {
                  "$ref": "#/components/schemas/SurfaceClaimState"
                }
              }
            }
          }
        }
      },
      "SurfaceClaimState": {
        "type": "string",
        "enum": [
          "unbound",
          "attached",
          "participating",
          "embedded-fallback",
          "degraded",
          "cross-boundary-refused",
          "quarantined",
          "detached"
        ]
      },
      "GateCheckSummary": {
        "type": "object",
        "required": ["name", "status", "score", "message"],
        "properties": {
          "name": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "score": {
            "type": ["string", "null"]
          },
          "message": {
            "type": "string"
          }
        }
      },
      "GateRunSummary": {
        "type": "object",
        "required": [
          "id",
          "result",
          "label",
          "score",
          "warning_count",
          "duration_seconds",
          "started_at",
          "new_warning_count",
          "changed_file_count",
          "checks"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "result": {
            "type": "string"
          },
          "label": {
            "type": "string"
          },
          "score": {
            "type": ["number", "null"]
          },
          "warning_count": {
            "type": "integer",
            "minimum": 0
          },
          "duration_seconds": {
            "type": ["number", "null"]
          },
          "started_at": {
            "type": ["string", "null"]
          },
          "new_warning_count": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "changed_file_count": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "checks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GateCheckSummary"
            }
          }
        }
      },
...
      "WarningSummary": {
        "type": "object",
        "required": [
          "id",
          "severity",
          "category",
          "message",
          "file_path",
          "age_label",
          "evidence_id",
          "rule",
          "line",
          "explanation",
          "matched_pattern",
          "evidence_excerpt"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "category": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "file_path": {
            "type": ["string", "null"]
          },
          "age_label": {
            "type": "string"
          },
          "evidence_id": {
            "type": "string"
          },
          "rule": {
            "type": "string"
          },
          "line": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "explanation": {
            "type": "string"
          },
          "matched_pattern": {
            "type": "string"
          },
          "evidence_excerpt": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EvidenceLine"
            }
          }
        }
      },
      "AttentionItem": {
        "type": "object",
        "required": ["title", "detail", "evidence_id"],
        "properties": {
          "title": {
            "type": "string"
          },
          "detail": {
            "type": "string"
          },
          "evidence_id": {
            "type": ["string", "null"]
          }
        }
      },
      "AffectedFile": {
        "type": "object",
        "required": [
          "path",
          "highest_severity",
          "warning_count",
          "first_seen",
          "last_seen",
          "warning_id"
        ],
        "properties": {
          "path": {
            "type": "string"
          },
          "highest_severity": {
            "type": "string"
          },
          "warning_count": {
            "type": "integer",
            "minimum": 0
          },
          "first_seen": {
            "type": "string"
          },
          "last_seen": {
            "type": "string"
          },
          "warning_id": {
            "type": "string"
          }
        }
      },
      "AssuranceSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "state",
          "reason",
          "generation",
          "last_full_scan",
          "scanned_files",
          "total_files"
        ],
        "properties": {
          "state": {
            "type": "string",
            "enum": ["clean", "stale", "pending", "running", "bounded", "unavailable", "unknown"]
          },
          "reason": {
            "type": ["string", "null"]
          },
          "generation": {
            "type": "integer",
            "minimum": 0
          },
          "last_full_scan": {
            "type": ["string", "null"]
          },
          "scanned_files": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "total_files": {
            "type": ["integer", "null"],
            "minimum": 0
          }
        }
      },
      "SaveTimeSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": ["state", "active", "failure_count"],
        "properties": {
          "state": {
            "type": "string",
            "enum": ["attached", "absent", "failed", "unknown"]
          },
          "active": {
            "type": "boolean"
          },
          "failure_count": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "DataGap": {
        "type": "object",
        "additionalProperties": false,
        "required": ["component", "reason"],
        "properties": {
          "component": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          }
        }
      },
      "ProtectionOverview": {
        "type": "object",
        "required": [
          "schema_version",
          "data_state",
          "source_message",
          "claim",
          "assurance",
          "save_time",
          "observed_at_unix",
          "latest_run",
          "recent_runs",
          "next_attention",
          "warnings_state",
          "warnings",
          "affected_files_state",
          "affected_files",
          "gaps"
        ],
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "anvil.dashboard.protection.v1"
          },
          "data_state": {
            "$ref": "#/components/schemas/DataState"
          },
          "source_message": {
            "type": "string"
          },
          "claim": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/ProtectionClaim"
              },
              {
                "type": "null"
              }
            ]
          },
          "assurance": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AssuranceSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "save_time": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/SaveTimeSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "observed_at_unix": {
            "type": ["integer", "null"],
            "minimum": 0
          },
          "latest_run": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/GateRunSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "recent_runs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/GateRunSummary"
            }
          },
          "next_attention": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/AttentionItem"
              },
              {
                "type": "null"
              }
            ]
          },
          "warnings_state": {
            "$ref": "#/components/schemas/DataState"
          },
          "warnings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/WarningSummary"
            }
          },
          "affected_files_state": {
            "$ref": "#/components/schemas/DataState"
          },
          "affected_files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AffectedFile"
            }
          },
          "gaps": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DataGap"
            }
          }
        }
      },
      "PatternSummary": {
        "type": "object",
        "required": [
          "id",
          "title",
          "family",
          "severity",
          "enabled",
          "instance_count",
          "description"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "family": {
            "type": "string"
          },
          "severity": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "instance_count": {
            "type": "integer",
            "minimum": 0
          },
          "description": {
            "type": "string"
          }
        }
      },
      "PatternCatalogue": {
        "type": "object",
        "required": ["schema_version", "data_state", "source_message", "patterns"],
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "anvil.dashboard.patterns.v1"
          },
          "data_state": {
            "$ref": "#/components/schemas/DataState"
          },
          "source_message": {
            "type": "string"
          },
          "patterns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatternSummary"
            }
          }
        }
      },
      "PlanSummary": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "scope", "title", "status", "progress"],
        "properties": {
          "id": {
            "type": "string"
          },
          "scope": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "progress": {
            "type": "string"
          }
        }
      },
      "PlanDetail": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "schema_version",
          "summary",
          "purpose",
          "actions_enabled",
          "action_message",
          "timeline"
        ],
        "properties": {
          "schema_version": {
            "type": "string",
            "const": "anvil.dashboard.plans.v1"
          },
          "summary": {
            "$ref": "#/components/schemas/PlanSummary"
          },
          "purpose": {
            "type": "string"
          },
          "actions_enabled": {
            "type": "boolean",
            "const": false
          },
          "action_message": {
            "type": "string"
          },
          "timeline": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanTimelineEntry"
            }
          }
        }
      },
      "PlanTimelineEntry": {
        "type": "object",
        "additionalProperties": false,
        "required": ["id", "title", "status", "validation_contract", "readiness"],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "validation_contract": {
            "type": ["string", "null"]
          },
          "readiness": {
            "type": "boolean"
          }
        }
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "required": ["code", "message"],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      }
    }
  }
}